requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
gunicorn==21.2.0
python-dotenv
psycopg2-binary